+++ pages/componentsC/table/table.vue
@@ -3,7 +3,7 @@
         <view class="u-page__item">
             <text class="u-page__item__title">基础使用</text>
             <u-table :data="tableData">
-                <u-table-column type="index" label="序号" />
+                <u-table-column type="index" label="序号"/>
                 <u-table-column prop="name" label="姓名" />
                 <u-table-column prop="address" label="地址" />
             </u-table>
@@ -21,7 +21,7 @@
         <view class="u-page__item">
             <text class="u-page__item__title">自定义列模板</text>
             <u-table :data="tableData" height="300">
-                <u-table-column type="index" label="序号" sortable />
+                <u-table-column type="index" label="序号大福大" sortable />
                 <u-table-column prop="name" label="姓名" sortable />
                 <u-table-column prop="status" label="状态" align="center">
                     <template v-slot="{ row, value, type }">