+++ uni_modules/uview-next/components/u-upload/u-upload.vue
@@ -157,6 +157,7 @@
 	 * @description 该组件用于上传图片场景
 	 * @tutorial https://uviewui.com/components/upload.html
 	 * @property {String}			accept				接受的文件类型, 可选值为all media image file video （默认 'image' ）
+	 * @property {String}			extension			根据文件拓展名过滤，每一项都不能是空字符串。默认不过滤。例如['.zip','.exe','.js']，不支持application/msword等类似值
 	 * @property {String | Array}	capture				图片或视频拾取模式，当accept为image类型时设置capture可选额外camera可以直接调起摄像头（默认 ['album', 'camera'] ）
 	 * @property {Boolean}			compressed			当accept为video时生效，是否压缩视频，默认为true（默认 true ）
 	 * @property {Boolean | Objcet} compressImage		当accept为image时生效，图片压缩参数，值为false不压缩
@@ -247,6 +248,7 @@
 				chooseFile(
 						Object.assign({
 							accept: this.accept,
+							extension: this.extension,
 							multiple: this.multiple,
 							capture: capture,
 							compressed: this.compressed,